requirements.txt
asgiref==3.11.0
Django==5.2.10
django-admin-honeypot==1.1.0
django-allauth==65.13.1
django-ckeditor==6.7.3
django-crispy-forms==2.5
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==3.1.2
django-modeltranslation==0.19.19
pillow==12.1.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
